fix(update): assert extracted binary is a regular file
